replaced the list for node objects by a dict {node.identifier : node} function get_index is not longer necessary node lookups can be done by self.nodes[nid] which is much more efficient then iterating through the list of nodes every time
I've made a performance test with a directory structure containing 2058 folders and 11255 files (13314 nodes incl. root). The program runs through the directories adding every folder and file to the tree. With the current implementation this process lasts about 35s. With the dict structure the program finishes after 0,5s.

The following function are currently disabled: paste, subtree as well as the module test